requirements.txt
pandas
numpy == 1.24.3
matplotlib
seaborn
scikit-learn
python-dotenv
pymongo
tensorflow
streamlit
pillow